workloads update
